requirements.txt
pdfplumber
python-pptx
pytesseract
Pillow
PyMuPDF
langchain
tqdm
sentence-transformers
chromadb
transformers
torch
numpy
langchain_community